requirements.txt
openai==1.7.0
langdetect==1.0.9
pyperclip==1.8.2
python-dateutil==2.8.2
markdown==3.5.1 